Telegram_Text/requirements.txt
pkg-resources==0.0.0
python-pytun==2.3.0
python-telegram==0.10.0
